requirements/dev.txt
-r common.txt

colorama==0.4.6
coverage==7.9.2
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1